chore(pkger): refactor Pkg tests out
